+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - **Modal**
   - Allow to prevent auto closing via `form.$onSubmitEnd = (e) => e.preventDefault()`
+  - Allow to prevent auto closing via `form.$onSubmit = (e) => e.preventDefault()`
 - **Controls**
   - Add css-animation for box-shadow/borders